Modifications of the emplacement of the code.
